requirements-ml.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
librosa>=0.10.1
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
timm>=0.9.12
transformers>=4.41.0
accelerate>=0.30.0
tqdm>=4.66.1
pyyaml>=6.0.1
matplotlib>=3.8.0
